updated sign up logic and server.js
