+++ routes/routes.js
@@ -1,6 +1,7 @@
 const express = require('express')
 const User = require('../models/user.model')
 const Transaction = require('../models/transaction.model')
+const GUser = require('./models/gusers.model')
 const bcrypt = require('bcrypt')
 require('dotenv').config()
 const jwt = require('jsonwebtoken')
@@ -14,6 +15,7 @@ const getTransaction = express.Router()
 const editTransaction = express.Router()
 const geteachTransaction = express.Router()
 const deleteTransaction = express.Router()
+const gusersRouter = express.Router()
 
 
 // SignUP route
@@ -32,11 +34,23 @@ signUpRouter.post("/signup", async (req, res) => {
         const hashedPassword = await bcrypt.hash(password, salt)
 
         let newUser = await User.create({ name, username, email, password: hashedPassword })
-        return res.status(200).json({ User: newUser })
+
+        // Create JWT token
+        const token = jwt.sign({ username: newUser.username }, process.env.SECRET_KEY);
+
+        // Store token in cookie
+        res.cookie('token', token, { httpOnly: true });
+
+        return res.status(200).json({
+            message: `Welcome, ${newUser.username}`,
+            user: newUser,
+            token
+        });
     } catch (err) {
-        return res.status(500).json({ success: false, message: err.message, });
+        return res.status(500).json({ success: false, message: err.message });
     }
 })
+
 
 // Login Route
 LoginRouter.post('/login', async (req, res) => {
@@ -178,9 +192,43 @@ deleteTransaction.delete('/delete/:id', async (req, res) => {
         console.error(err);
         res.status(500).json({ error: 'Something went wrong' });
     }
-
-
 })
 
+gusersRouter.post('/gusers', (req, res) => {
+    const userData = req.body;
+    const { email } = userData;
 
-module.exports = { signUpRouter, LoginRouter, LogoutRouter, transactionRouter, getTransaction, editTransaction, geteachTransaction, deleteTransaction }
+    // Check if user already exists
+    GUser.findOne({ email })
+        .then(existingUser => {
+            if (existingUser) {
+                // User already exists, return existing user data
+                const token = jwt.sign({ userId: existingUser._id }, process.env.SECRET_KEY, { expiresIn: '1h' });
+                res.status(200).json({ user: existingUser, token });
+            } else {
+                // User doesn't exist, create a new user
+                const newUser = new GUser(userData);
+                newUser.save()
+                    .then(savedUser => {
+                        console.log('User data saved to MongoDB:', savedUser);
+                        const token = jwt.sign({ userId: savedUser._id }, 'your_secret_key', { expiresIn: '1h' });
+                        res.status(200).json({ user: savedUser, token });
+                    })
+                    .catch(err => {
+                        console.error('Error saving user data to MongoDB:', err);
+                        res.status(500).send('Error saving user data to MongoDB');
+                    });
+            }
+        })
+        .catch(err => {
+            console.error('Error finding user in MongoDB:', err);
+            res.status(500).send('Error finding user in MongoDB');
+        });
+});
+
+
+
+
+
+
+module.exports = { signUpRouter, LoginRouter, LogoutRouter, transactionRouter, getTransaction, editTransaction, geteachTransaction, deleteTransaction, gusersRouter }
